+++ packages/e2e-tests/lib/runTestApp.ts
@@ -1,14 +1,10 @@
-/* eslint-disable no-console */
-
-import * as fs from 'fs-extra';
+import * as fs from 'fs';
+import * as fsExtra from 'fs-extra';
 import * as path from 'path';
 
 import { buildApp } from './buildApp';
-import { TMP_DIR } from './constants';
 import { testApp } from './testApp';
 import type { Env, RecipeInstance, RecipeTestResult } from './types';
-
-let tmpDirCount = 0;
 
 // This should never throw, we always return a result here
 export async function buildAndTestApp(
@@ -18,9 +14,13 @@ export async function buildAndTestApp(
   const { recipe, portModulo, portGap } = recipeInstance;
   const recipeDirname = path.dirname(recipe.path);
 
-  const targetDir = path.join(TMP_DIR, `${recipe.testApplicationName}-${tmpDirCount++}`);
+  const tmpFolder = path.join(__dirname, '..', 'tmp');
+  await fs.promises.mkdir(tmpFolder, { recursive: true });
+  const targetDir = await fs.promises.mkdtemp(
+    path.join(tmpFolder, `${recipeInstance.recipe.testApplicationName}-${Date.now()}-`),
+  );
 
-  await fs.copy(recipeDirname, targetDir);
+  await fsExtra.copy(recipeDirname, targetDir);
 
   const env: Env = {
     ...envVarsToInject,
@@ -31,7 +31,7 @@ export async function buildAndTestApp(
   try {
     await buildApp(targetDir, recipeInstance, env);
   } catch (error) {
-    await fs.remove(targetDir);
+    await fsExtra.remove(targetDir);
 
     return {
       ...recipeInstance,
@@ -42,15 +42,17 @@ export async function buildAndTestApp(
   }
 
   // This cannot throw, we always return a result here
-  const results = await testApp(targetDir, recipeInstance, env);
-
-  // Cleanup
-  await fs.remove(targetDir);
-
-  return {
-    ...recipeInstance,
-    buildFailed: false,
-    testFailed: results.some(result => result.result !== 'PASS'),
-    tests: results,
-  };
+  return testApp(targetDir, recipeInstance, env)
+    .finally(() => {
+      // Cleanup
+      void fsExtra.remove(targetDir);
+    })
+    .then(results => {
+      return {
+        ...recipeInstance,
+        buildFailed: false,
+        testFailed: results.some(result => result.result !== 'PASS'),
+        tests: results,
+      };
+    });
 }